Adding unit test for submessage issue.
The two messages are supposed to be the same size going across the wire, but the fail message has two arrays in it, with the second one overridding the first one.

The only difference between the two messages is that the fail array calls an initialize method inside it's default constructor.

updated the pom file to only call the sub message test.